scr/bot/requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
faiss-cpu==1.8.0.post1
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.1
langchain-community==0.2.16
langchain-core==0.2.39
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.117
magic-filter==1.0.12
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yandex-chain==0.0.9
yarl==1.11.1
